gen/requirements.txt
django_redis
openai
celery
mysqlclient
